+++ lib/Data/Time/LocalTime/Internal/TimeZone.hs
@@ -57,21 +57,26 @@ minutesToTimeZone m = TimeZone m False ""
 hoursToTimeZone :: Int -> TimeZone
 hoursToTimeZone i = minutesToTimeZone (60 * i)
 
-showT :: PadOption -> Int -> String
-showT opt t = showPaddedNum opt ((div t 60) * 100 + (mod t 60))
+showT :: Bool -> PadOption -> Int -> String
+showT False opt t = showPaddedNum opt ((div t 60) * 100 + (mod t 60))
+showT True opt t = let
+    opt' = case opt of
+        NoPad -> NoPad
+        Pad i c -> Pad (max 0 $ i - 3) c
+    in showPaddedNum opt' (div t 60) ++ ":" ++ show2 (mod t 60)
 
-timeZoneOffsetString'' :: PadOption -> TimeZone -> String
-timeZoneOffsetString'' opt (TimeZone t _ _) | t < 0 = '-':(showT opt (negate t))
-timeZoneOffsetString'' opt (TimeZone t _ _) = '+':(showT opt t)
+timeZoneOffsetString'' :: Bool -> PadOption -> TimeZone -> String
+timeZoneOffsetString'' colon opt (TimeZone t _ _) | t < 0 = '-':(showT colon opt (negate t))
+timeZoneOffsetString'' colon opt (TimeZone t _ _) = '+':(showT colon opt t)
 
 -- | Text representing the offset of this timezone, such as \"-0800\" or \"+0400\" (like @%z@ in formatTime), with arbitrary padding.
 timeZoneOffsetString' :: Maybe Char -> TimeZone -> String
-timeZoneOffsetString' Nothing = timeZoneOffsetString'' NoPad
-timeZoneOffsetString' (Just c) = timeZoneOffsetString'' $ Pad 4 c
+timeZoneOffsetString' Nothing = timeZoneOffsetString'' False NoPad
+timeZoneOffsetString' (Just c) = timeZoneOffsetString'' False $ Pad 4 c
 
 -- | Text representing the offset of this timezone, such as \"-0800\" or \"+0400\" (like @%z@ in formatTime).
 timeZoneOffsetString :: TimeZone -> String
-timeZoneOffsetString = timeZoneOffsetString'' (Pad 4 '0')
+timeZoneOffsetString = timeZoneOffsetString'' False (Pad 4 '0')
 
 instance Show TimeZone where
     show zone@(TimeZone _ _ "") = timeZoneOffsetString zone